+++ go/internal/latex/assemble.go
@@ -145,8 +145,8 @@ func (r *Runner) bookSessionTools(proj, buildDir string, compile *CompileTexTool
 		&GrepTool{Root: buildDir, AltRoots: []AltRoot{{Label: "project", Dir: r.projectRoot()}}},
 		&WorkBashTool{Root: buildDir, Mounts: r.sessionMounts(kindBook, buildDir), MaxOutput: r.cfg.BashMaxOutput(), Sandbox: r.cfg.BashSandboxEnabled(), Log: r.log, Tid: 1},
 		compile,
-		&ViewPDFTool{Mounts: r.sessionMounts(kindBook, buildDir), Comp: r.comp},
-		&ViewImageTool{Root: buildDir},
+		&ViewPDFTool{Mounts: r.sessionMounts(kindBook, buildDir), Comp: r.comp, SoftMax: r.cfg.ViewPDFMax(), WarnRatio: r.cfg.ViewWarnRatio()},
+		&ViewImageTool{Root: buildDir, SoftMax: r.cfg.ViewImageMax(), WarnRatio: r.cfg.ViewWarnRatio()},
 		&ListFontsTool{FontsDir: r.cfg.Paths.Fonts},
 		submit,
 	}